+++ src/rendering/render_minimap.c
@@ -1,0 +1,1 @@
+#include "cub3d.h"
